# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SystemGatewayAPI",
    "version": "1.0"
  },
  "paths": {
    "/api/Applications": {
      "post": {
        "tags": [
          "Applications"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Application"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Application"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Application"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Applications"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Applications/{ApplicationName}/Version/{VersionId}": {
      "put": {
        "tags": [
          "Applications"
        ],
        "parameters": [
          {
            "name": "ApplicationName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "VersionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModuleVersion"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ModuleVersion"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ModuleVersion"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Applications"
        ],
        "parameters": [
          {
            "name": "ApplicationName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "VersionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Applications/{ApplicationName}": {
      "delete": {
        "tags": [
          "Applications"
        ],
        "parameters": [
          {
            "name": "ApplicationName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Applications"
        ],
        "parameters": [
          {
            "name": "ApplicationName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Authentication/Authenticate/{UserType}": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "parameters": [
          {
            "name": "userType",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UserType"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthenticateInputDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthenticateInputDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AuthenticateInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Authentication/Information/{Token}": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SessionData"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionData"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionData"
                }
              }
            }
          }
        }
      }
    },
    "/api/Authentication/{Token}/Modules": {
      "delete": {
        "tags": [
          "Authentication"
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SessionData"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionData"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionData"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Application": {
        "type": "object",
        "properties": {
          "applicationName": {
            "type": "string",
            "nullable": true
          },
          "versions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModuleVersion"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AuthenticateInputDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DataPoint": {
        "type": "object",
        "properties": {
          "sectionName": {
            "type": "string",
            "nullable": true
          },
          "isDataEditable": {
            "type": "boolean"
          },
          "content": {
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ModuleSnapshot": {
        "type": "object",
        "properties": {
          "moduleId": {
            "type": "string",
            "format": "uuid"
          },
          "checksum": {
            "type": "string",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "ModuleVersion": {
        "type": "object",
        "properties": {
          "versionId": {
            "type": "string",
            "nullable": true
          },
          "applicationName": {
            "type": "string",
            "nullable": true
          },
          "dataStructure": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataPoint"
            },
            "nullable": true
          },
          "htmlCard": {
            "type": "string",
            "nullable": true
          },
          "htmlDashboard": {
            "type": "string",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "checksum": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "SessionData": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "userType": {
            "$ref": "#/components/schemas/UserType"
          },
          "expires": {
            "type": "string",
            "format": "date-time"
          },
          "isExpired": {
            "type": "boolean",
            "readOnly": true
          },
          "moduleSnapshots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModuleSnapshot"
            },
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UserType": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      }
    }
  }
}